Using multiple trackers in groups to detect presence
How to configure multiple trackers (in this example, Owntracks, Nmap, and iBeacons) in a group so that, if just one tracker reports a 'home' condition, the user's device is considered to be 'at home.' This example differs from others in which just one 'away' condition' from a tracker causes the user's device to be considered 'away.' This configuration would be useful if your phone's Owntracks position tends to periodically drift outside of your Home zone. The additional trackers in the group prevent false 'away' states.